requirements.txt
eel>=0.16.0
pyserial>=3.5
numpy>=1.24.0
Pillow>=10.0.0
# pycromanager – install separately; requires Micro-Manager running on port 4827
# pycromanager>=0.27.0
